WIP: Add ByteBufUtils
This adds ByteBufUtils, which are an evolution of ByteBufWriteUtils,
covering reads of Uint types as well as writing using various approaches
to nullness.

We also migrate all uint reading tasks to these utility methods, fixing
deprecation warnings in the process.

+++ bgp/parser-impl/src/main/java/org/opendaylight/protocol/bgp/parser/impl/message/open/LlGracefulCapabilityHandler.java
@@ -85,7 +85,7 @@ public final class LlGracefulCapabilityHandler implements CapabilityParser, Capa
                     .setAfi(afi)
                     .setSafi(safi)
                     .setAfiFlags(new Tables.AfiFlags(Boolean.valueOf(afiFlags == AFI_FLAG_FORWARDING_STATE)))
-                    .setLongLivedStaleTime(Long.valueOf(staleTime))
+                    .setLongLivedStaleTime(Uint32.valueOf(staleTime))
                     .build();
             tables.add(table);
         }
